hf-requirements.txt
torch
transformers>=4.40
pillow
requests

